Move fingerprinting to base
