+++ src/sentry/utils/performance_issues/base.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
+import hashlib
 import random
 from abc import ABC, abstractmethod
 from datetime import timedelta
 from enum import Enum
-from typing import Any, Dict
+from typing import Any, Dict, List
 
 from sentry import options
 from sentry.eventstore.models import Event
@@ -151,3 +152,28 @@ def get_url_from_span(span: Span) -> str:
         url = url.get("pathname") or ""
 
     return url
+
+
+def fingerprint_spans(spans: List[Span]):
+    span_hashes = []
+    for span in spans:
+        hash = span.get("hash", "") or ""
+        span_hashes.append(str(hash))
+    joined_hashes = "-".join(span_hashes)
+    return hashlib.sha1(joined_hashes.encode("utf8")).hexdigest()
+
+
+# Creates a stable fingerprint given the same span details using sha1.
+def fingerprint_span(span: Span):
+    op = span.get("op", None)
+    description = span.get("description", None)
+    if not description or not op:
+        return None
+
+    signature = (str(op) + str(description)).encode("utf-8")
+    full_fingerprint = hashlib.sha1(signature).hexdigest()
+    fingerprint = full_fingerprint[
+        :20
+    ]  # 80 bits. Not a cryptographic usage, we don't need all of the sha1 for collision detection
+
+    return fingerprint
